+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Updated syntax highlighting spec up to ([d92b40a](https://github.com/jtroo/kanata/tree/d92b40a))
+* Added support for `deflocalkeys-winiov2`
 
 ### 0.11.0